// silap-colmirwcl/src/app/modules/silapapi/model/versamentoProvincia.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Provincia } from './provincia';
import { VersamentoPvCob } from './versamentoPvCob';
import { VersamentoPvConvenzione } from './versamentoPvConvenzione';
import { VersamentoPvEsonero } from './versamentoPvEsonero';
import { VersamentoPvGgExtraFestivi } from './versamentoPvGgExtraFestivi';
import { VersamentoPvPeriodo } from './versamentoPvPeriodo';
import { VersamentoPvRicIn } from './versamentoPvRicIn';
import { VersamentoPvSospensione } from './versamentoPvSospensione';

export interface VersamentoProvincia { 
    codUserAggiorn?: string;
    codUserInserim?: string;
    dAggiorn?: Date;
    dFestaPatronale?: Date;
    dInserim?: Date;
    esoTVersamentoPvCobs?: Array<VersamentoPvCob>;
    esoTVersamentoPvConvenziones?: Array<VersamentoPvConvenzione>;
    esoTVersamentoPvEsoneros?: Array<VersamentoPvEsonero>;
    esoTVersamentoPvGgExtraFestivis?: Array<VersamentoPvGgExtraFestivi>;
    esoTVersamentoPvPeriodos?: Array<VersamentoPvPeriodo>;
    esoTVersamentoPvRicIns?: Array<VersamentoPvRicIn>;
    esoTVersamentoPvSospensiones?: Array<VersamentoPvSospensione>;
    idEsoTVersamento?: number;
    idEsoTVersamentoProvincia?: number;
    importoAutomatico?: number;
    nTimestamp?: number;
    numGgLavorativiSettimanali?: number;
    silapDProvincia?: Provincia;
}