// silap-colmirwcl/src/app/modules/silapapi/model/versamentoPvProspetto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Provincia } from './provincia';

export interface VersamentoPvProspetto { 
    baseComputoProvinciale?: number;
    catCompensazioneDisabili?: string;
    codUserAggiorn?: string;
    codUserInserim?: string;
    dAggiorn?: Date;
    dInserim?: Date;
    idEsoTVersamentoPvProspetto?: number;
    nTimestamp?: number;
    numDisabiliInForza?: number;
    numEsoneratiAutocertificati?: number;
    numSoggettiCompensatiDisabili?: number;
    quotaRiservaDisabili?: number;
    silapDProvincia?: Provincia;
}